+++ src/machine.c
@@ -227,18 +227,18 @@ bool ghost_machine_bigendian()
     return (endiantest[0] == 0);
 }
 
-size_t ghost_machine_alignment()
+int ghost_machine_alignment()
 {
 #ifdef GHOST_HAVE_MIC
-    size_t alignment = 64;
+    int alignment = 64;
 #elif defined(GHOST_HAVE_AVX2)
-    size_t alignment = 32;
+    int alignment = 32;
 #elif defined(GHOST_HAVE_AVX)
-    size_t alignment = 32;
+    int alignment = 32;
 #elif defined(GHOST_HAVE_SSE)
-    size_t alignment = 16;
+    int alignment = 16;
 #else
-    size_t alignment = 8;
+    int alignment = 8;
 #endif
    return alignment; 
 }
